feat: implement dynamic component rendering and refactor interface mapping
