+++ src/app/components/interfaceMapping.tsx
@@ -1,6 +1,6 @@
-import { Checkbox, ToggleSwitchProps } from '../components/checkbox/checkbox'
-import { RadioBar, RadioBarProps } from '../components/radios/radio-bar'
-import { TextFieldProps, TextInput } from '../components/text-field/text-field'
+import { Checkbox, ToggleSwitchProps } from './checkbox/checkbox'
+import { RadioBar, RadioBarProps } from './radios/radio-bar'
+import { TextFieldProps, TextInput } from './text-field/text-field'
 
 export interface Option<T> {
     optionProps: T
